Found the logic that makes the calculator "work"
Inside the "equalTo" function we call the current value of the input content and save it inside
a constant, then, we check which operator it contains, so when the operator is found inside of the array,
we cut it in two parts and use the operator that was found.

So with this information, we can call the function operate with these three arguments.

Now i need to finish correctly the way to make the result pop up when a operator is put in for a second time, and then,
show the results in the display of the calculator.